+++ internal/registry/owner_other.go
@@ -1,9 +1,0 @@
-//go:build windows
-
-package registry
-
-import "os"
-
-func checkOwner(p string, fi os.FileInfo) error {
-	return nil
-}